# Use one version of transitively vendored deps