… pex/vendor/__init__.py
… pex/vendor/__main__.py
… pex/vendor/_vendored/appdirs/appdirs-1.4.3.dist-info/DESCRIPTION.rst
… pex/vendor/_vendored/appdirs/appdirs-1.4.3.dist-info/INSTALLER
… pex/vendor/_vendored/appdirs/appdirs-1.4.3.dist-info/METADATA
… pex/vendor/_vendored/appdirs/appdirs-1.4.3.dist-info/WHEEL
… pex/vendor/_vendored/appdirs/appdirs-1.4.3.dist-info/metadata.json
… pex/vendor/_vendored/appdirs/appdirs-1.4.3.dist-info/top_level.txt
… pex/vendor/_vendored/appdirs/appdirs.py
… pex/vendor/_vendored/appdirs/constraints.txt
… pex/vendor/_vendored/packaging/packaging/requirements.py
… pex/vendor/_vendored/pip/pip-20.0.dev0.dist-info/entry_points.txt
… pex/vendor/_vendored/setuptools/pkg_resources/__init__.py
… pex/vendor/_vendored/setuptools/pkg_resources/_vendor/packaging/markers.py
… pex/vendor/_vendored/setuptools/pkg_resources/_vendor/packaging/requirements.py
… pex/vendor/_vendored/setuptools/setuptools/__init__.py
… pex/vendor/_vendored/setuptools/setuptools/_vendor/packaging/markers.py
… pex/vendor/_vendored/setuptools/setuptools/_vendor/packaging/requirements.py
… pex/vendor/_vendored/setuptools/setuptools/command/alias.py
… pex/vendor/_vendored/setuptools/setuptools/command/bdist_egg.py
… pex/vendor/_vendored/setuptools/setuptools/command/build_ext.py
… pex/vendor/_vendored/setuptools/setuptools/command/build_py.py
… pex/vendor/_vendored/setuptools/setuptools/command/develop.py
… pex/vendor/_vendored/setuptools/setuptools/command/easy_install.py
… pex/vendor/_vendored/setuptools/setuptools/command/egg_info.py
… pex/vendor/_vendored/setuptools/setuptools/command/py36compat.py
… pex/vendor/_vendored/setuptools/setuptools/command/rotate.py
… pex/vendor/_vendored/setuptools/setuptools/command/sdist.py
… pex/vendor/_vendored/setuptools/setuptools/command/setopt.py
… pex/vendor/_vendored/setuptools/setuptools/command/test.py