This is kind of hacky. It probably shouldn't be a hard-coded list in
`__main__.py`.

It's also not guaranteed to work - it relies on us picking good versions
to vendor which are compatibile with all of the transitive Avendored
versions.

We also don't vendor all of the transitive deps, because the rest don't
seem to be used by our codepaths. This is a bit fragile.

Finally, the last rewrite, `pex.third_party.six` ->
`pex.vendor._vendored.six.six` shouldn't be necessary. If the approach
seems reasonable, I'll dig into why it's needed to work, and resolve
that.

Performance difference:
Before:
```
pex$ mkdir -p trivial && echo 'print("Hello")' > trivial/trivial.py && python3.6 -m pex -D trivial -m trivial -o trivial.pex && multitime -n 50 python3.6 ./trivial.pex >/dev/null
===> multitime results
1: python3.6 ./trivial.pex
            Mean        Std.Dev.    Min         Median      Max 
real        0.466       0.024       0.434       0.466       0.594    
user        0.409       0.022       0.380       0.408       0.527          
sys         0.054       0.002       0.049       0.054       0.063 
```

After:
```
pex$ mkdir -p trivial && echo 'print("Hello")' > trivial/trivial.py && python3.6 -m pex -D trivial -m trivial -o trivial.pex && multitime -n 50 python3.6 ./trivial.pex >/dev/null
===> multitime results
1: python3.6 ./trivial.pex
            Mean        Std.Dev.    Min         Median      Max 
real        0.405       0.025       0.368       0.402       0.519    
user        0.355       0.023       0.325       0.352       0.457    
sys         0.046       0.003       0.041       0.046       0.057  
```

Relates to #930