@illicitonion 2021-09-08
> Stale?

Yeah realistically if I revive this at some point it'll be effectively from scratch :) Closing for now